+++ src/main/java/com/github/ontio/network/rest/RestClient.java
@@ -318,7 +318,7 @@ public class RestClient extends AbstractConnector {
     }
 
     @Override
-    public String getSideChainData(int sideChainID) throws ConnectorException, IOException {
+    public String getSideChainData(long sideChainID) throws ConnectorException, IOException {
         return null;
     }
